Group components by packages
