+++ core/src/main/java/com/jaspersoft/android/sdk/service/InfoCacheManagerImpl.java
@@ -26,6 +26,8 @@ package com.jaspersoft.android.sdk.service;
 
 import com.jaspersoft.android.sdk.service.data.server.ServerInfo;
 import com.jaspersoft.android.sdk.service.exception.ServiceException;
+import com.jaspersoft.android.sdk.service.info.InfoCache;
+import com.jaspersoft.android.sdk.service.info.InfoCacheManager;
 import com.jaspersoft.android.sdk.service.server.ServerInfoService;
 import org.jetbrains.annotations.TestOnly;
