+++ core/src/main/java/com/jaspersoft/android/sdk/service/report/ReportExportUseCase.java
@@ -31,8 +31,8 @@ import com.jaspersoft.android.sdk.network.entity.execution.ExecutionStatus;
 import com.jaspersoft.android.sdk.network.entity.export.ExportExecutionDescriptor;
 import com.jaspersoft.android.sdk.network.entity.export.ExportOutputResource;
 import com.jaspersoft.android.sdk.network.entity.export.OutputResource;
-import com.jaspersoft.android.sdk.service.Call;
-import com.jaspersoft.android.sdk.service.CallExecutor;
+import com.jaspersoft.android.sdk.service.call.Call;
+import com.jaspersoft.android.sdk.service.call.CallExecutor;
 import com.jaspersoft.android.sdk.service.data.report.ReportOutput;
 import com.jaspersoft.android.sdk.service.data.report.ResourceOutput;
 import com.jaspersoft.android.sdk.service.exception.ServiceException;
