+++ core/src/main/java/com/jaspersoft/android/sdk/service/repository/SearchTaskImpl.java
@@ -25,8 +25,8 @@
 package com.jaspersoft.android.sdk.service.repository;
 
 import com.jaspersoft.android.sdk.network.RepositoryRestApi;
-import com.jaspersoft.android.sdk.service.CallExecutor;
-import com.jaspersoft.android.sdk.service.InfoCacheManager;
+import com.jaspersoft.android.sdk.service.call.CallExecutor;
+import com.jaspersoft.android.sdk.service.info.InfoCacheManager;
 import com.jaspersoft.android.sdk.service.data.repository.Resource;
 import com.jaspersoft.android.sdk.service.exception.ServiceException;
 import org.jetbrains.annotations.NotNull;
