+++ core/src/main/java/com/jaspersoft/android/sdk/service/repository/SearchUseCase.java
@@ -27,9 +27,9 @@ package com.jaspersoft.android.sdk.service.repository;
 import com.jaspersoft.android.sdk.network.HttpException;
 import com.jaspersoft.android.sdk.network.RepositoryRestApi;
 import com.jaspersoft.android.sdk.network.entity.resource.ResourceSearchResult;
-import com.jaspersoft.android.sdk.service.Call;
-import com.jaspersoft.android.sdk.service.CallExecutor;
-import com.jaspersoft.android.sdk.service.InfoCacheManager;
+import com.jaspersoft.android.sdk.service.call.Call;
+import com.jaspersoft.android.sdk.service.call.CallExecutor;
+import com.jaspersoft.android.sdk.service.info.InfoCacheManager;
 import com.jaspersoft.android.sdk.service.data.repository.Resource;
 import com.jaspersoft.android.sdk.service.data.repository.SearchResult;
 import com.jaspersoft.android.sdk.service.exception.ServiceException;
