+++ core/src/main/java/com/jaspersoft/android/sdk/service/token/TokenCache.java
@@ -22,27 +22,14 @@
  * <http://www.gnu.org/licenses/lgpl>.
  */
 
-package com.jaspersoft.android.sdk.service;
+package com.jaspersoft.android.sdk.service.token;
 
 /**
  * @author Tom Koptel
  * @since 2.0
  */
-public final class InMemoryTokenCache implements TokenCache {
-    private String mCache;
-
-    @Override
-    public String get() {
-        return mCache;
-    }
-
-    @Override
-    public void put(String token) {
-        mCache = token;
-    }
-
-    @Override
-    public void evict() {
-        mCache = null;
-    }
+public interface TokenCache {
+    String get();
+    void put(String token);
+    void evict();
 }
