+++ core/src/test/java/com/jaspersoft/android/sdk/service/repository/RepositoryServiceTest.java
@@ -26,7 +26,7 @@ package com.jaspersoft.android.sdk.service.repository;
 
 import com.jaspersoft.android.sdk.network.RepositoryRestApi;
 import com.jaspersoft.android.sdk.service.FakeCallExecutor;
-import com.jaspersoft.android.sdk.service.InfoCacheManager;
+import com.jaspersoft.android.sdk.service.info.InfoCacheManager;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
