+++ core/src/test/java/com/jaspersoft/android/sdk/service/repository/SearchTaskImplTest.java
@@ -26,7 +26,7 @@ package com.jaspersoft.android.sdk.service.repository;
 
 import com.jaspersoft.android.sdk.network.RepositoryRestApi;
 import com.jaspersoft.android.sdk.service.FakeCallExecutor;
-import com.jaspersoft.android.sdk.service.InfoCacheManager;
+import com.jaspersoft.android.sdk.service.info.InfoCacheManager;
 import com.jaspersoft.android.sdk.service.data.repository.Resource;
 import com.jaspersoft.android.sdk.service.data.server.ServerInfo;
 import org.junit.Before;
